remove kernel package
