+++ internal/validators.go
@@ -15,7 +15,33 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package kernel
+package internal
 
-type TransactionListener interface {
+import (
+	"regexp"
+
+	"github.com/apache/rocketmq-client-go/rlog"
+)
+
+const (
+	_ValidPattern       = "^[%|a-zA-Z0-9_-]+$"
+	_CharacterMaxLength = 255
+)
+
+var (
+	_Pattern, _ = regexp.Compile("_ValidPattern")
+)
+
+func ValidateGroup(group string) {
+	if group == "" {
+		rlog.Fatal("consumerGroup is empty")
+	}
+
+	//if !_Pattern.Match([]byte(group)) {
+	//	rlog.Fatalf("the specified group[%s] contains illegal characters, allowing only %s", group, _ValidPattern)
+	//}
+
+	if len(group) > _CharacterMaxLength {
+		rlog.Fatal("the specified group is longer than group max length 255.")
+	}
 }
